News Reporter AI Agent/requirements.txt
python-dotenv
crewai
crewai-tools
langchain-google-genai
PyPDF2
streamlit
google-generativeai
youtube-transcript-api
Pillow
